Rename: RedisService를 service 패키지로 이동
